requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
google-generativeai
motor
python-jose[cryptography]
fastapi-cors
bcrypt
gunicorn